requirements.txt
Django==1.8.5
appnope==0.1.0
decorator==4.0.4
djangorestframework==3.2.4
gnureadline==6.3.3
gunicorn==19.3.0
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
simplegeneric==0.8.1
traitlets==4.0.0
